Adding picture upload on AddEventActivity. Show location progress on AddEventActivity. Working on androidTest. Create dev/uat app keys

+++ app/src/main/java/com/timappweb/timapp/config/EventStatusManager.java
@@ -33,7 +33,7 @@ public class EventStatusManager {
     private static EventStatusManager _instance = null;
     private static Event currentEvent;
 
-    public static boolean isCurrentEvent(int eventId) {
+    public static boolean isCurrentEvent(long eventId) {
         return currentEvent != null && currentEvent.getRemoteId() == eventId;
     }
